// src/app/core/models/models.ts
export interface Category {
  _id: string;
  description: string; // Include the 'description' property here if it exists
  Libelle: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}